+++ examples/queries.rs
@@ -27,17 +27,13 @@ enum QueryError {
 #[derive(Clone, PartialEq, Eq, Hash, Debug)]
 enum QueryValue {
     UserName(String),
-}
-
-#[derive(Clone, PartialEq, Eq, Hash, Debug)]
-enum MutationValue {
-    UserUpdated(usize),
+    UserAge(u8),
 }
 
 async fn fetch_user(keys: Vec<QueryKeys>) -> QueryResult<QueryValue, QueryError> {
     if let Some(QueryKeys::User(id)) = keys.first() {
-        println!("Fetching user {id}");
-        sleep(Duration::from_millis(1000)).await;
+        println!("Fetching name of user {id}");
+        sleep(Duration::from_millis(650)).await;
         match id {
             0 => Ok(QueryValue::UserName("Marc".to_string())),
             _ => Err(QueryError::UserNotFound(*id)),
@@ -48,33 +44,31 @@ async fn fetch_user(keys: Vec<QueryKeys>) -> QueryResult<QueryValue, QueryError>
     }
 }
 
-async fn update_user((id, _name): (usize, String)) -> MutationResult<MutationValue, QueryError> {
-    println!("Mutating user");
-    sleep(Duration::from_millis(1000)).await;
-    Ok(MutationValue::UserUpdated(id)).into()
+async fn fetch_user_age(keys: Vec<QueryKeys>) -> QueryResult<QueryValue, QueryError> {
+    if let Some(QueryKeys::User(id)) = keys.first() {
+        println!("Fetching age of user {id}");
+        sleep(Duration::from_millis(1000)).await;
+        match id {
+            0 => Ok(QueryValue::UserAge(0)),
+            _ => Err(QueryError::UserNotFound(*id)),
+        }
+        .into()
+    } else {
+        QueryResult::Err(QueryError::Unknown)
+    }
 }
 
 #[allow(non_snake_case)]
 #[component]
 fn User(cx: Scope, id: usize) -> Element {
-    let value = use_query(cx, || vec![QueryKeys::User(*id)], fetch_user);
-    let mutate = use_mutation(cx, update_user);
-
-    let onclick = |_| {
-        mutate.mutate((0, "Not Marc".to_string()));
-    };
+    let user_name = use_query(cx, move || vec![QueryKeys::User(*id)], fetch_user);
+    let user_age = use_query(cx, move || vec![QueryKeys::User(*id)], fetch_user_age);
 
     println!("Showing user {id}");
 
     render!(
-        p { "{value.result().value():?}" }
-        button { onclick: onclick,
-            if mutate.result().is_loading() {
-              "Loading..."
-           } else {
-               "Fake mutation"
-           }
-        }
+        p { "{user_name.result().value():?}" }
+        p { "{user_age.result().value():?}" }
     )
 }
 
@@ -82,11 +76,9 @@ fn app(cx: Scope) -> Element {
     use_init_query_client::<QueryValue, QueryError, QueryKeys>(cx);
     let client = use_query_client::<QueryValue, QueryError, QueryKeys>(cx);
 
-    let refresh = move |_| client.invalidate_query(QueryKeys::User(0));
+    let refresh = |_| client.invalidate_query(QueryKeys::User(0));
 
     render!(
-        User { id: 0 }
-        User { id: 0 }
         User { id: 0 }
         User { id: 0 }
         button { onclick: refresh, label { "Refresh" } }
